+++ deltafi-ui/src/composables/useDeltaFilesQueryBuilder.ts
@@ -22,7 +22,7 @@ import useGraphQL from './useGraphQL'
 export default function useDeltaFilesQueryBuilder() {
   const { response, queryGraphQL } = useGraphQL();
 
-  const getDeltaFileSearchData = (startDateISOString: String, endDateISOString: String, offSet: Number, perPage: Number, sortBy: string, sortDirection: string, fileName?: string, stageName?: string, actionName?: string, flowName?: string, egressFlowName?: string, egressed?: Boolean, filtered?: Boolean, domain?: string, metadata?: Array<Record<string, string>>, ingressBytesMin?: Number, ingressBytesMax?: Number, totalBytesMin?: Number, totalBytesMax?: Number,testMode?: Boolean,requeueMin?:Number) => {
+  const getDeltaFileSearchData = (startDateISOString: String, endDateISOString: String, offSet: Number, perPage: Number, sortBy: string, sortDirection: string, fileName?: string, stageName?: string, actionName?: string, flowName?: Array<string>, egressFlowName?: Array<string>, egressed?: Boolean, filtered?: Boolean, domain?: string, metadata?: Array<Record<string, string>>, ingressBytesMin?: Number, ingressBytesMax?: Number, totalBytesMin?: Number, totalBytesMax?: Number,testMode?: Boolean,requeueMin?:Number) => {
     const query = {
       deltaFiles: {
         __args: {
@@ -33,10 +33,10 @@ export default function useDeltaFilesQueryBuilder() {
             filtered: filtered,
             testMode: testMode,
             sourceInfo: {
-              flow: flowName,
+              ingressFlows: flowName ? flowName : [],
               filename: fileName
             },
-            egressFlows: egressFlowName ? [egressFlowName] : [],
+            egressFlows: egressFlowName ? egressFlowName : [],
             stage: stageName ? new EnumType(stageName) : null,
             actions: actionName,
             modifiedAfter: startDateISOString,